[TOOLS-2414] Fix selected tab behavior in issue details
Description: Comment tab is not open in subtask after opening a card in another lane

Add the 'restamp' property according API References:
When true, elements will be removed from DOM and discarded when if becomes false and re-created
and added back to the DOM when if becomes true. By default, stamped elements will be hidden but
left in the DOM when if becomes false, which is generally results in better performance.